Move FingerPrintClientHello into u_fingerprinter.go along with Fingerprinter struct which allows for specifying options
